backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
sentence-transformers
requests
python-dotenv